requirements.txt
djangorestframework==3.16.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
pydantic-settings==2.11.0
PyJWT==2.10.1
dependency-injector==4.48.2